Reorganize project to reduce unnecessary dependency
