+++ director/origin_api.go
@@ -39,8 +39,9 @@ import (
 
 	"github.com/pelicanplatform/pelican/config"
 	"github.com/pelicanplatform/pelican/param"
+	"github.com/pelicanplatform/pelican/server_utils"
+	"github.com/pelicanplatform/pelican/token"
 	"github.com/pelicanplatform/pelican/token_scopes"
-	"github.com/pelicanplatform/pelican/utils"
 )
 
 type (
@@ -118,12 +119,12 @@ func checkNamespaceStatus(prefix string, registryWebUrlStr string) (bool, error)
 // see if the entity is authorized to advertise an origin for the
 // namespace
 func VerifyAdvertiseToken(ctx context.Context, token, namespace string) (bool, error) {
-	issuerUrl, err := GetNSIssuerURL(namespace)
+	issuerUrl, err := server_utils.GetNSIssuerURL(namespace)
 	if err != nil {
 		return false, err
 	}
 
-	keyLoc, err := GetJWKSURLFromIssuerURL(issuerUrl)
+	keyLoc, err := server_utils.GetJWKSURLFromIssuerURL(issuerUrl)
 	if err != nil {
 		return false, err
 	}
@@ -202,16 +203,16 @@ func VerifyAdvertiseToken(ctx context.Context, token, namespace string) (bool, e
 // Verify that a token received is a valid token from director
 func VerifyDirectorTestReportToken(strToken string) (bool, error) {
 	directorURL := param.Federation_DirectorUrl.GetString()
-	token, err := jwt.Parse([]byte(strToken), jwt.WithVerify(false))
+	parsedToken, err := jwt.Parse([]byte(strToken), jwt.WithVerify(false))
 	if err != nil {
 		return false, err
 	}
 
-	if directorURL != token.Issuer() {
+	if directorURL != parsedToken.Issuer() {
 		return false, errors.Errorf("Token issuer is not a director")
 	}
 
-	key, err := utils.LoadDirectorPublicKey()
+	key, err := token.LoadDirectorPublicKey()
 	if err != nil {
 		return false, err
 	}
@@ -238,75 +239,4 @@ func VerifyDirectorTestReportToken(strToken string) (bool, error) {
 		}
 	}
 	return false, nil
-}
-
-// For a given prefix, get the prefix's issuer URL, where we consider that the openid endpoint
-// we use to look up a key location. Note that this is NOT the same as the issuer key -- to
-// find that, follow openid-style discovery using the issuer URL as a base.
-func GetNSIssuerURL(prefix string) (string, error) {
-	if prefix == "" || !strings.HasPrefix(prefix, "/") {
-		return "", errors.New(fmt.Sprintf("the prefix \"%s\" is invalid", prefix))
-	}
-	registryUrlStr := param.Federation_RegistryUrl.GetString()
-	if registryUrlStr == "" {
-		return "", errors.New("federation registry URL is not set and was not discovered")
-	}
-	registryUrl, err := url.Parse(registryUrlStr)
-	if err != nil {
-		return "", err
-	}
-
-	registryUrl.Path, err = url.JoinPath(registryUrl.Path, "api", "v1.0", "registry", prefix)
-
-	if err != nil {
-		return "", errors.Wrapf(err, "failed to construct openid-configuration lookup URL for prefix %s", prefix)
-	}
-	return registryUrl.String(), nil
-}
-
-// Given an issuer url, lookup the JWKS URL from the openid-configuration
-// For example, if the issuer URL is https://registry.com:8446/api/v1.0/registry/test-namespace,
-// this function will return the key indicated by the openid-configuration JSON hosted at
-// https://registry.com:8446/api/v1.0/registry/test-namespace/.well-known/openid-configuration.
-func GetJWKSURLFromIssuerURL(issuerUrl string) (string, error) {
-	// Get/parse the openid-configuration JSON to lookup key location
-	issOpenIDUrl, err := url.Parse(issuerUrl)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to parse issuer URL")
-	}
-	issOpenIDUrl.Path, _ = url.JoinPath(issOpenIDUrl.Path, ".well-known", "openid-configuration")
-
-	client := &http.Client{Transport: config.GetTransport()}
-	openIDCfg, err := client.Get(issOpenIDUrl.String())
-	if err != nil {
-		return "", errors.Wrapf(err, "failed to lookup openid-configuration for issuer %s", issuerUrl)
-	}
-	defer openIDCfg.Body.Close()
-
-	// If we hit an old registry, it may not have the openid-configuration. In that case, we fallback to the old
-	// behavior of looking for the key directly at the issuer URL.
-	if openIDCfg.StatusCode == http.StatusNotFound {
-		oldKeyLoc, err := url.JoinPath(issuerUrl, ".well-known", "issuer.jwks")
-		if err != nil {
-			return "", errors.Wrapf(err, "failed to construct key lookup URL for issuer %s", issuerUrl)
-		}
-		return oldKeyLoc, nil
-	}
-
-	body, err := io.ReadAll(openIDCfg.Body)
-	if err != nil {
-		return "", errors.Wrapf(err, "failed to read response body from %s", issuerUrl)
-	}
-
-	var openIDCfgMap map[string]string
-	err = json.Unmarshal(body, &openIDCfgMap)
-	if err != nil {
-		return "", errors.Wrapf(err, "failed to unmarshal openid-configuration for issuer %s", issuerUrl)
-	}
-
-	if keyLoc, ok := openIDCfgMap["jwks_uri"]; ok {
-		return keyLoc, nil
-	} else {
-		return "", errors.New(fmt.Sprintf("no key found in openid-configuration for issuer %s", issuerUrl))
-	}
 }
