+++ server_utils/test_file_transfer.go
@@ -1,6 +1,6 @@
 /***************************************************************
  *
- * Copyright (C) 2023, Pelican Project, Morgridge Institute for Research
+ * Copyright (C) 2024, Pelican Project, Morgridge Institute for Research
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you
  * may not use this file except in compliance with the License.  You may
@@ -19,7 +19,7 @@
 // This is a utility file that provides a TestFileTransferImpl struct with a `RunTests` function
 // to allow any Pelican server to issue a file transfer test to a XRootD server
 
-package utils
+package server_utils
 
 import (
 	"bytes"
@@ -31,6 +31,7 @@ import (
 
 	"github.com/pelicanplatform/pelican/config"
 	"github.com/pelicanplatform/pelican/param"
+	"github.com/pelicanplatform/pelican/token"
 	"github.com/pkg/errors"
 )
 
@@ -76,8 +77,8 @@ func (t TestFileTransferImpl) generateFileTestScitoken() (string, error) {
 		return "", errors.New("Failed to create token: Invalid iss, Server_ExternalWebUrl is empty")
 	}
 
-	fTestTokenCfg := TokenConfig{
-		TokenProfile: WLCG,
+	fTestTokenCfg := token.TokenConfig{
+		TokenProfile: token.WLCG,
 		Lifetime:     time.Minute,
 		Issuer:       issuerUrl,
 		Audience:     t.audiences,
