+++ token/token_verify.go
@@ -1,6 +1,6 @@
 /***************************************************************
  *
- * Copyright (C) 2023, Pelican Project, Morgridge Institute for Research
+ * Copyright (C) 2024, Pelican Project, Morgridge Institute for Research
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you
  * may not use this file except in compliance with the License.  You may
@@ -15,7 +15,8 @@
  * limitations under the License.
  *
  ***************************************************************/
-package utils
+
+package token
 
 import (
 	"context"
